Adicao de novas features
Tela de confirmacao e timer

+++ componentes/cadastrarPeguntasScreen.js
@@ -33,7 +33,7 @@ export default function CadastrarPerguntaScreen() {
 
   async function carregaDados() {
     await carregarTemas();
-    await carregarPerguntas();    
+    await carregarPerguntas();
   }
 
 
@@ -82,36 +82,35 @@ export default function CadastrarPerguntaScreen() {
   };
 
   async function salvarPergunta() {
-  if (!temaSelecionado || !pergunta.trim() || alternativas.some(a => !a.trim())) {
-    Alert.alert('Erro', 'Preencha todos os campos');
-    return;
-  }
-
-  try {
-    const perguntaId = await addPergunta(pergunta, temaSelecionado, respostaCorreta);
-
-    if (!perguntaId) {
-      throw new Error('Erro ao inserir pergunta no banco.');
+    if (!temaSelecionado || !pergunta.trim() || alternativas.some(a => !a.trim())) {
+      Alert.alert('Erro', 'Preencha todos os campos');
+      return;
     }
 
-    console.log("antes de inserir alternativassss")
-    for (let i = 0; i < alternativas.length; i++) {
-      const sucesso = await addAlternativa(perguntaId, alternativas[i], i);
-      if (!sucesso) {
-        throw new Error(`Erro ao salvar a alternativa ${i + 1}`);
+    try {
+      const perguntaId = await addPergunta(pergunta, temaSelecionado, respostaCorreta);
+
+      if (!perguntaId) {
+        throw new Error('Erro ao inserir pergunta no banco.');
       }
+
+      console.log("antes de inserir alternativassss")
+      for (let i = 0; i < alternativas.length; i++) {
+        const sucesso = await addAlternativa(perguntaId, alternativas[i], i);
+        if (!sucesso) {
+          throw new Error(`Erro ao salvar a alternativa ${i + 1}`);
+        }
+      }
+
+      Alert.alert('Sucesso', 'Pergunta salva!');
+      resetForm();
+      carregarPerguntas();
+
+    } catch (error) {
+      console.error(error);
+      Alert.alert('Erro', error.message);
     }
-
-    Alert.alert('Sucesso', 'Pergunta salva!');
-    resetForm();
-    carregarPerguntas();
-
-  } catch (error) {
-    console.error(error);
-    Alert.alert('Erro', error.message);
   }
-}
-
 
   async function iniciarEdicao(pergunta) {
     setEditandoId(pergunta.id);
@@ -124,46 +123,78 @@ export default function CadastrarPerguntaScreen() {
     console.log(vetor);
 
     setAlternativasEditadas(vetor);
+
   }
 
-  async function salvarEdicao() {
+  function confirmaEdicao() {
+
+    Alert.alert("Aviso", "Deseja mesmo editar a pergunta?", [
+      { text: "Sim", onPress: () => salvarEdicao(true) },
+      { text: "Não", onPress: () => salvarEdicao(false) }
+    ]);
+
+  }
+
+  async function salvarEdicao(isConfirmed) {
     if (!temaEditado || !perguntaEditada.trim() || alternativasEditadas.some(a => !a.trim())) {
       Alert.alert('Erro', 'Preencha todos os campos');
       return;
     }
 
-    const sucesso = await updatePergunta(editandoId, perguntaEditada, temaEditado, respostaCorretaEditada);
-    if (sucesso) {
-      // Apaga alternativas antigas
-      const altAntigas = await getAlternativas(editandoId);
-      for (const alt of altAntigas) {
-        await deleteAlternativa(alt.id);
-      }
-      // Adiciona alternativas novas
-      for (let i = 0; i < alternativasEditadas.length; i++) {
-        await addAlternativa(editandoId, alternativasEditadas[i], i);
-      }
+    if (isConfirmed) {
 
-      Alert.alert('Sucesso', 'Pergunta atualizada!');
-      resetEditForm();
-      carregarPerguntas();
+      const sucesso = await updatePergunta(editandoId, perguntaEditada, temaEditado, respostaCorretaEditada);
+      if (sucesso) {
+        // Apaga alternativas antigas
+        const altAntigas = await getAlternativas(editandoId);
+        for (const alt of altAntigas) {
+          await deleteAlternativa(alt.id);
+        }
+        // Adiciona alternativas novas
+        for (let i = 0; i < alternativasEditadas.length; i++) {
+          await addAlternativa(editandoId, alternativasEditadas[i], i);
+        }
+
+        Alert.alert('Sucesso', 'Pergunta atualizada!');
+        resetEditForm();
+        carregarPerguntas();
+      } else {
+        Alert.alert('Erro', 'Erro ao atualizar pergunta');
+      }
     } else {
-      Alert.alert('Erro', 'Erro ao atualizar pergunta');
+      Alert.alert('Cancelado', 'Pergunta não atualizada');
     }
+
   }
 
-  async function removerPergunta(id) {
-    const sucesso = await deletePergunta(id);
-    if (sucesso) {
-      const alt = await getAlternativas(id);
-      for (const a of alt) {
-        await deleteAlternativa(a.id);
+  function confirmaExclusao(id) {
+
+    Alert.alert("Aviso", "Deseja mesmo remover a pergunta?", [
+      { text: "Sim", onPress: () => removerPergunta(id, true) },
+      { text: "Não", onPress: () => removerPergunta(id, false) }
+    ]);
+
+  }
+
+  async function removerPergunta(id, isConfirmed) {
+
+    if (isConfirmed) {
+      const sucesso = await deletePergunta(id);
+      if (sucesso) {
+        const alt = await getAlternativas(id);
+        for (const a of alt) {
+          await deleteAlternativa(a.id);
+        }
+        Alert.alert('Sucesso', 'Pergunta removida!');
+        carregarPerguntas();
+      } else {
+        Alert.alert('Erro', 'Erro ao remover pergunta');
       }
-      Alert.alert('Sucesso', 'Pergunta removida!');
-      carregarPerguntas();
+
     } else {
-      Alert.alert('Erro', 'Erro ao remover pergunta');
+      Alert.alert('Cancelado', 'Pergunta não removida');
     }
+
   }
 
   return (
@@ -237,7 +268,7 @@ export default function CadastrarPerguntaScreen() {
 
       {editandoId ? (
         <View style={styles.buttonsRow}>
-          <TouchableOpacity style={[styles.button, styles.saveButton]} onPress={salvarEdicao}>
+          <TouchableOpacity style={[styles.button, styles.saveButton]} onPress={confirmaEdicao}>
             <Text style={styles.buttonText}>Salvar Alterações</Text>
           </TouchableOpacity>
           <TouchableOpacity style={[styles.button, styles.cancelButton]} onPress={resetEditForm}>
@@ -260,7 +291,7 @@ export default function CadastrarPerguntaScreen() {
             <TouchableOpacity onPress={() => iniciarEdicao(item)} style={styles.iconBtn}>
               <Image source={iconEdit} style={styles.icon} />
             </TouchableOpacity>
-            <TouchableOpacity onPress={() => removerPergunta(item.id)} style={styles.iconBtn}>
+            <TouchableOpacity onPress={() => confirmaExclusao(item.id)} style={styles.iconBtn}>
               <Image source={iconDelete} style={styles.icon} />
             </TouchableOpacity>
           </View>
@@ -274,13 +305,13 @@ export default function CadastrarPerguntaScreen() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#D5F5E3',  
+    backgroundColor: '#D5F5E3',
     padding: 20,
   },
   title: {
     fontWeight: 'bold',
     fontSize: 22,
-    color: '#2E5137', 
+    color: '#2E5137',
     marginBottom: 15,
     textAlign: 'center',
   },
@@ -317,7 +348,7 @@ const styles = StyleSheet.create({
     marginBottom: 10,
   },
   saveButton: {
-    backgroundColor: '#7BB661', 
+    backgroundColor: '#7BB661',
   },
   cancelButton: {
     backgroundColor: '#A9A9A9',
